requirements.txt
transformers==4.30.2
torch==2.0.1
numpy==1.24.3
pandas==2.0.2
spacy==3.5.0
python-docx==0.8.11
PyPDF2==3.0.1
pydantic==1.10.11
nltk==3.8.1
scikit-learn==1.2.2
requests==2.31.0
